Change method of stopping click propagation in confirmClick
The directive was stopping click propagation by declaring it as a
terminal directive and when the user confirmed the dialog, it would
evaluate the ng-click content. This blocks any other directive applied
to the button to process.

Now, when the user cancels the dialog, only the click event is stopped
from propagating, and other directives can execute without problems.