Replace API error handling with directive
The api-click directive executes the action passed to the api-click
attribute and parses its result, saving the final state in the object
passed to the api-click-state attribute.

All the error handling implemented in controllers were replaced with
this directive. In forms, the submit button is now a regular button
that on click executes the action that was specified in ng-submit.