Merge busyWhen with apiClick to use isolate scope
`busyWhen` requires an isolate scope, due to the `when` property that it
changes in the current scope, otherwise in case there are multiple `busyWhen`
uses in the same view, they might end altering the same variable.

As multiple directives can't require an isolate scope in the same element,
`busyWhen` is merged into `apiClick`, as an optional attribute.