Fix reboot on the container
/sbin/reboot would shutdown the container instead of rebooting it.
This patch exposes the Docker UNIX socket so that we can call
docker commands inside the container. With this we call
`docker restart webui` on reboot.

This patch also ignores errors on calling reboot and return the
request early, signaling it succeeded. Sometimes reboot can shutdown
the machine before the server can respond the request, as it happens
with the docker container, signaling it failed when it has not.

Change-Id: I135142d31b560df3341883d15cba458e28875f5c
Reviewed-on: https://review.gerrithub.io/409106
Reviewed-by: Thiago Figueredo Cardoso <thiago.figueredo@cesar.org.br>
Tested-by: Thiago Figueredo Cardoso <thiago.figueredo@cesar.org.br>