devices: Refactor add/remove monitoring
In order to enable devices registered on fog to be shown on dashboard
this patch refactors the watching operations to create a connection
with the message broker and add the new devices properly. Also, this
patch renames and avoid repetition on code to keep compatibility with
devices that are added through a DBus interface.